+++ saleor/graphql/page/types.py
@@ -93,9 +93,8 @@ class PageType(CountableDjangoObjectType):
     @staticmethod
     @permission_required(PagePermissions.MANAGE_PAGES)
     def resolve_has_pages(root: models.PageType, info, **kwargs):
-        def check_if_has_pages(pages):
-            return bool(pages)
-
         return (
-            PagesByPageTypeIdLoader(info.context).load(root.pk).then(check_if_has_pages)
+            PagesByPageTypeIdLoader(info.context)
+            .load(root.pk)
+            .then(lambda pages: bool(pages))
         )